Fixes for 64 bit Windows build: - fix CMake files to choose correct Simbody library names - added many casts to avoid warnings due to size_t and ptrdiff_t values being put into int variables (size_t is unsigned long long for 64 bit builds; ptrdiff_t is long long). (There are now no warnings at all, at least for VS10.)
I also added an advanced variable in the CMake build to limit the number of parallel compilations that are performed; default is still no limit.

Problems:
- testProbe is failing (both 32 and 64 bit builds)
- testOptimizationExample fails in 64 bit build
- testWrapping crashes my machine in 64 bit mode -- but first its memory use goes to over 2.5GB!
- testExampleMainRuns launches the visualizer and hangs until the visualizer is manually closed

git-svn-id: file:///dev/shm/opensim/Trunk@8004 fc29be9c-2cb0-420d-b520-73dbd18add6e

+++ OpenSim/Actuators/Millard2012AccelerationMuscle.cpp
@@ -388,8 +388,8 @@ Array<std::string> Millard2012AccelerationMuscle::getStateVariableNames() const
 SimTK::SystemYIndex Millard2012AccelerationMuscle::
     getStateVariableSystemIndex(const std::string &stateVariableName) const
 {
-    unsigned int start = stateVariableName.find(".");
-	unsigned int end = stateVariableName.length();
+    unsigned start = (unsigned)stateVariableName.find(".");
+	unsigned end = (unsigned)stateVariableName.length();
 	
 	if(start == end)
 		return ModelComponent::getStateVariableSystemIndex(stateVariableName);
