+++ OpenSim/Simulation/SimbodyEngine/Test/testConstraints.cpp
@@ -108,7 +108,7 @@ public:
 	}
 	virtual double calcDerivative(const std::vector<int>& derivComponents, const SimTK::Vector& x) const
 	{
-	   int nd = derivComponents.size();
+	   int nd = (int)derivComponents.size();
 	   if (nd < 1)
 		   return SimTK::NaN;
 	   if (derivComponents[0] == 0){
